+++ go/chat/commands/flip.go
@@ -30,27 +30,25 @@ func (s *Flip) Execute(ctx context.Context, uid gregor1.UID, convID chat1.Conver
 
 func (s *Flip) Preview(ctx context.Context, uid gregor1.UID, convID chat1.ConversationID, text string) {
 	defer s.Trace(ctx, func() error { return nil }, "Preview")()
-	if !s.Match(ctx, text) || text == "/flip" {
+	if !s.Match(ctx, text) {
 		s.getChatUI().ChatCommandMarkdown(ctx, convID, "")
 		return
 	}
 	cur := s.G().CoinFlipManager.DescribeFlipText(ctx, text)
 	usage := fmt.Sprintf(`
-	Current Flip: %s
-
-    Example commands:
+	Example commands: %s
         /flip          coin flip
         /flip 6        roll a 6-sided die (1..6)
         /flip 10..20   pick a number 10 to 20 (inclusive)
         /flip a,b,c,d  shuffle some options and pick where to eat or whom to wrestle
-        /flip cards    shuffle and deal a deck
-
-    And for a quick game of face-up poker:
-        /flip cards 5 @user1 @user2 @user3
-            (shuffle a deck and deal 5 cards to 3 different people
+        /flip cards    shuffle and deal a deck %s
+    And for a quick game of face-up poker: %s		/flip cards 5 @user1 @user2 @user3
+            (shuffle a deck and deal 5 cards to 3 different people%s
 
     The blog post announcing this feature and how it works:
-        https://keybase.io/coin-flipping
-`, cur)
+		https://keybase.io/coin-flipping
+
+	_Current Flip_: %s
+`, "```", "```", "```", "```", cur)
 	s.getChatUI().ChatCommandMarkdown(ctx, convID, usage)
 }
