+++ go/chat/commands/source.go
@@ -71,7 +71,6 @@ func (s *Source) makeBuiltins() {
 	common := []types.ConversationCommand{
 		cmds[cmdCollapse],
 		cmds[cmdExpand],
-		cmds[cmdFlip],
 		cmds[cmdGiphy],
 		cmds[cmdHide],
 		cmds[cmdMe],
@@ -79,6 +78,9 @@ func (s *Source) makeBuiltins() {
 		cmds[cmdMute],
 		cmds[cmdShrug],
 		cmds[cmdUnhide],
+	}
+	if s.isAdmin() {
+		common = append(common, cmds[cmdFlip])
 	}
 	s.builtins = make(map[chat1.ConversationBuiltinCommandTyp][]types.ConversationCommand)
 	s.builtins[chat1.ConversationBuiltinCommandTyp_ADHOC] = common
