+++ shared/chat/conversation/input-area/normal/types.js
@@ -24,6 +24,7 @@ type CommonProps = {|
   onRequestScrollDown: () => void,
   onRequestScrollUp: () => void,
   onSubmit: (text: string) => void,
+  showCommandMarkdown: boolean,
   showWalletsIcon: boolean, // used on mobile to determine placeholder
 
   editText: string,
