+++ shared/chat/conversation/normal/index.types.js.flow
@@ -18,6 +18,5 @@ export type Props = {
   onShowTracker: (username: string) => void,
   onToggleInfoPanel: () => void,
   threadLoadedOffline: boolean,
-  showCommandMarkdown: boolean,
   showGiphySearch: boolean,
 }
